pyproject.toml
[tool.poetry]
name = "hof-videos"
version = "0.1.0"
description = "Personal hall of fame videos."
authors = ["alpden550 <alpden550@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
python-dotenv = "^0.13.0"
psycopg2 = "^2.8.5"
django-widget-tweaks = "^1.4.8"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
autopep8 = "^1.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
